feature: impl all utils and functionality in lib for now
