+++ examples/hello_lib.rs
@@ -1,8 +1,8 @@
-use core;
+use core::server::Server;
 
 fn main() {
   let host = "127.0.0.1";
   let port = 4000;
 
-  core::server::Server::listen(host, port);
+  Server::listen(host, port);
 }
